# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "goact API",
        "title": "goact RESTful API",
        "contact": {
            "name": "API Support",
            "email": "support@mck-p.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/license/mit/"
        },
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/communities": {
            "post": {
                "description": "This will create a new Community and assign the creator as the only memberof that community",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/vnd.api+json"
                ],
                "tags": [
                    "Communities"
                ],
                "summary": "Creates a new Community",
                "operationId": "CreateCommunity",
                "parameters": [
                    {
                        "description": "New Community Information",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.CreateCommunityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.SuccessResponse-data_Community"
                        }
                    }
                }
            }
        },
        "/api/v1/messages/groups": {
            "get": {
                "produces": [
                    "application/vnd.api+json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Retrieves the Message Groups that a User has access to",
                "operationId": "GetUserMessageGroups",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.SuccessResponse-server_GetGroupsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse-server_GenericError"
                        }
                    }
                }
            },
            "post": {
                "description": "This will create a new Message Group and assign the creator as the only authorized user of that message group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/vnd.api+json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Creates a new Message Group",
                "operationId": "CreateMessageGroup",
                "parameters": [
                    {
                        "description": "Message Group Information",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.MessageGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.SuccessResponse-data_MessageGroup"
                        }
                    }
                }
            }
        },
        "/api/v1/messages/groups/{:id}": {
            "get": {
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.SuccessResponse-server_GetMessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse-server_GenericError"
                        }
                    }
                }
            }
        },
        "/api/v1/users/:id/messages": {
            "get": {
                "produces": [
                    "application/vnd.api+json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Retrieves the last N number of messages, given some offset",
                "operationId": "GetMessages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit for pagination",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.SuccessResponse-server_GetMessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse-server_GenericError"
                        }
                    }
                }
            }
        },
        "/api/v1/users/external-id/{externalid}": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "produces": [
                    "application/vnd.api+json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Returns the Goact User based on their external ID",
                "operationId": "GetUserByExternalId",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.SuccessResponse-data_User"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse-server_GenericError"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "produces": [
                    "application/vnd.api+json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Returns the Goact User based on their internal ID",
                "operationId": "GetUserById",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.SuccessResponse-data_User"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse-server_GenericError"
                        }
                    }
                }
            }
        },
        "/api/v1/weather": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "produces": [
                    "application/vnd.api+json"
                ],
                "tags": [
                    "Weather"
                ],
                "summary": "Gets the last-checked weather conditions",
                "operationId": "GetWeather",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.SuccessResponse-data_ForecastQueryUpstreamResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse-server_GenericError"
                        }
                    }
                }
            }
        },
        "/api/v1/webhooks": {
            "post": {
                "description": "Handles incoming webhooks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/vnd.api+json"
                ],
                "tags": [
                    "Internal"
                ],
                "summary": "Handle Webhook",
                "operationId": "WebhookIngestion",
                "parameters": [
                    {
                        "description": "Webhook payload",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.SuccessResponse-any"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "data.AstroCondition": {
            "type": "object",
            "properties": {
                "is_sun_up": {
                    "type": "number"
                },
                "moon_illumination": {
                    "type": "number"
                },
                "moon_phase": {
                    "type": "string"
                },
                "moonrise": {
                    "type": "string"
                },
                "moonset": {
                    "type": "string"
                },
                "sunrise": {
                    "type": "string"
                },
                "sunset": {
                    "type": "string"
                }
            }
        },
        "data.Community": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "data.ForecastQueryUpstreamResponse": {
            "type": "object",
            "properties": {
                "current": {
                    "$ref": "#/definitions/data.CurrentWeatherUpstreamResponse"
                },
                "forecast": {
                    "type": "object",
                    "properties": {
                        "forecastday": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.ForecastUpstreamResponse"
                            }
                        }
                    }
                }
            }
        },
        "data.ForecastUpstreamResponse": {
            "type": "object",
            "properties": {
                "astro": {
                    "$ref": "#/definitions/data.AstroCondition"
                },
                "day": {
                    "$ref": "#/definitions/data.DayWeatherUpstreamResponse"
                },
                "hour": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/data.HourWeatherUpstreamResponse"
                    }
                }
            }
        },
...
        "data.MessageGroup": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "data.User": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "avatarUrl": {
                    "type": "string"
                },
                "externalid": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "server.CreateCommunityRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "server.ErrorResponse-server_GenericError": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/server.GenericError"
                },
                "includes": {
                    "type": "array",
                    "items": {}
                },
                "links": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/server.Link"
                    }
                },
                "metadata": {}
            }
        },
...
        "server.GenericError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "server.GetGroupsResponse": {
            "type": "object",
            "properties": {
                "groups": {
                    "type": "array",
                    "items": {}
                }
            }
        },
        "server.GetMessageResponse": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {}
                }
            }
        },
...
        "server.Link": {
            "type": "object",
            "properties": {
                "href": {
                    "type": "string"
                },
                "hreflang": {
                    "type": "string"
                },
                "link": {
                    "$ref": "#/definitions/server.Link"
                },
                "meta": {},
                "rel": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "server.MessageGroupRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "server.SuccessResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "includes": {
                    "type": "array",
                    "items": {}
                },
                "links": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/server.Link"
                    }
                },
                "metadata": {}
            }
        },
        "server.SuccessResponse-data_Community": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/data.Community"
                },
                "includes": {
                    "type": "array",
                    "items": {}
                },
                "links": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/server.Link"
                    }
                },
                "metadata": {}
            }
        },
        "server.SuccessResponse-data_ForecastQueryUpstreamResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/data.ForecastQueryUpstreamResponse"
                },
                "includes": {
                    "type": "array",
                    "items": {}
                },
                "links": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/server.Link"
                    }
                },
                "metadata": {}
            }
        },
        "server.SuccessResponse-data_MessageGroup": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/data.MessageGroup"
                },
                "includes": {
                    "type": "array",
                    "items": {}
                },
                "links": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/server.Link"
                    }
                },
                "metadata": {}
            }
        },
        "server.SuccessResponse-data_User": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/data.User"
                },
                "includes": {
                    "type": "array",
                    "items": {}
                },
                "links": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/server.Link"
                    }
                },
                "metadata": {}
            }
        },
        "server.SuccessResponse-server_GetGroupsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/server.GetGroupsResponse"
                },
                "includes": {
                    "type": "array",
                    "items": {}
                },
                "links": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/server.Link"
                    }
                },
                "metadata": {}
            }
        },
        "server.SuccessResponse-server_GetMessageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/server.GetMessageResponse"
                },
                "includes": {
                    "type": "array",
                    "items": {}
                },
                "links": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/server.Link"
                    }
                },
                "metadata": {}
            }
        },
...
    },
    "securityDefinitions": {
        "BearerToken": {
            "description": "JWT for authentication. Inclue Bearer",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}